调整 - 翻译器重构（停止、继续、自动保存）

+++ Module_Folders/Translator.py
@@ -5,29 +5,26 @@ import json
 import threading
 import concurrent.futures
 
-from PyQt5.QtCore import QTimer
-from PyQt5.QtCore import QObject
-
 from Base.AiNieeBase import AiNieeBase
 from Module_Folders.File_Reader.File1 import File_Reader
 from Module_Folders.Cache_Manager.Cache import Cache_Manager
 from Module_Folders.File_Outputer.File2 import File_Outputter
+from Module_Folders.Request_Limiter.Request_limit import Request_Limiter
 from Module_Folders.TranslatorTask import TranslatorTask
 
-
 # 翻译器
-class Translator(QObject, AiNieeBase):
+class Translator(AiNieeBase):
 
-    # 缓存文件保存周期
-    CACHE_FILE_SAVE_TICK = 3000
+    # 缓存文件保存周期（毫秒）
+    CACHE_FILE_SAVE_INTERVAL = 3000
 
-    def __init__(self, configurator, plugin_manager, request_limiter):
+    def __init__(self, configurator, plugin_manager):
         super().__init__()
 
         # 初始化
         self.configurator = configurator
         self.plugin_manager = plugin_manager
-        self.request_limiter = request_limiter
+        self.request_limiter = Request_Limiter(configurator)
 
         # 线程锁
         self.lock = threading.Lock()
@@ -35,27 +32,62 @@ class Translator(QObject, AiNieeBase):
         self.cache_file_lock = threading.Lock()
 
         # 注册事件
+        self.subscribe(self.EVENT.TRANSLATION_STOP, self.translation_stop)
         self.subscribe(self.EVENT.TRANSLATION_START, self.translation_start)
+        self.subscribe(self.EVENT.TRANSLATION_CONTINUE_CHECK, self.translation_continue_check)
 
         # 定时器
-        self.timer = QTimer()
-        self.timer.timeout.connect(self.save_cache_file_tick)
-        self.timer.start(self.CACHE_FILE_SAVE_TICK)
+        threading.Thread(target = self.save_cache_file_tick).start()
+
+    # 翻译停止事件
+    def translation_stop(self, event: int, data: dict):
+        # 设置运行状态为停止中
+        self.configurator.Running_status = self.STATUS.STOPING
+
+        def target():
+            while True:
+                time.sleep(0.5)
+                if len([t for t in threading.enumerate() if "translator" in t.name]) == 0:
+                    self.configurator.Running_status = self.STATUS.IDLE
+                    self.emit(self.EVENT.TRANSLATION_STOP_DONE, {})
+                    break
+
+        threading.Thread(target = target).start()
 
     # 翻译开始事件
     def translation_start(self, event: int, data: dict):
         threading.Thread(
             target = self.translation_start_target,
-            args = ()
+            args = (data.get("translation_continue"),),
         ).start()
 
+    # 翻译状态检查事件
+    def translation_continue_check(self, event: int, data: dict):
+        translation_continue = False
+        try:
+            cache = self.load_cache_file(f"{self.configurator.label_output_path}/cache/AinieeCacheData.json")
+            untranslated_text_line_count, _ = Cache_Manager.count_and_update_translation_status_0_2(
+                self,
+                cache,
+            )
+            translation_continue = untranslated_text_line_count > 0
+        except Exception:
+            self.error("缓存文件读取失败 ...", e)
+        finally:
+            self.emit(self.EVENT.TRANSLATION_CONTINUE_CHECK_DONE, {
+                "translation_continue" : translation_continue,
+            })
+
     # 实际的翻译流程
-    def translation_start_target(self):
+    def translation_start_target(self, translation_continue):
         # 设置翻译状态为正在翻译状态
         self.configurator.Running_status = self.STATUS.TRANSLATION
 
-        # 初始化数据
-        self.data = {
+        # 读取配置文件
+        self.configurator.load_config_file()
+
+        # 根据是否继续翻译载入项目数据
+        default_data = {
             "status": "翻译中",
             "start_time": time.time(),
             "total_line": 0,
@@ -69,22 +101,38 @@ class Translator(QObject, AiNieeBase):
             "task": 0,
         }
 
+        if translation_continue:
+            # 读取文本
+            try:
+                self.configurator.cache_list = self.load_cache_file(f"{self.configurator.label_output_path}/cache/AinieeCacheData.json")
+            except Exception as e:
+                self.error("翻译项目数据载入失败 ... ", e)
+                return {}
+
+            # 初始化数据
+            if len(self.configurator.cache_list) == 0:
+                self.data = default_data
+            else:
+                self.data = self.configurator.cache_list[0].get("data", {})
+                self.data["status"] = "翻译中"
+                self.data["start_time"] = time.time() - self.data.get("time", 0)
+        else:
+            # 读取文本
+            try:
+                self.configurator.cache_list = File_Reader.read_files(
+                    self,
+                    self.configurator.translation_project,
+                    self.configurator.label_input_path,
+                )
+            except Exception as e:
+                self.error("翻译项目数据载入失败 ... ", e)
+                return {}
+
+            # 初始化数据
+            self.data = default_data
+
         # 更新翻译进度
         self.emit(self.EVENT.TRANSLATION_UPDATE, self.data)
-
-        # 读取配置文件
-        self.configurator.load_config_file()
-
-        # 读取输入文件
-        try:
-            self.configurator.cache_list = File_Reader.read_files(
-                self,
-                self.configurator.translation_project,
-                self.configurator.label_input_path,
-            )
-        except Exception as e:
-            self.error("读取原文文件失败，请检查项目类型是否设置正确，输入文件夹是否混杂其他非必要文件！", e)
-            return
 
         # 调用插件，进行文本过滤
         self.plugin_manager.broadcast_event(
@@ -110,6 +158,10 @@ class Translator(QObject, AiNieeBase):
         # 开始循环
         time.sleep(3)
         for current_round in range(self.configurator.round_limit + 1):
+            # 检测是否需要停止任务
+            if self.configurator.Running_status == self.STATUS.STOPING:
+                return {}
+
             # 译前准备工作
             task_current_round, line_current_round = self.translation_prepare(current_round)
 
@@ -127,8 +179,8 @@ class Translator(QObject, AiNieeBase):
                 self.print("")
                 break
 
-            # 第一轮时记录开始总行数
-            if current_round == 0:
+            # 第一轮时且不是继续翻译时，记录总行数
+            if current_round == 0 and translation_continue == False:
                 self.data["total_line"] = line_current_round
 
             # 输出开始翻译的日志
@@ -167,6 +219,9 @@ class Translator(QObject, AiNieeBase):
                         ).start
                     )
                     future.add_done_callback(lambda future: self.task_done_callback(future, executor))
+
+        # 等待可能存在的缓存文件写入请求处理完毕
+        time.sleep(self.CACHE_FILE_SAVE_INTERVAL/1000)
 
         # 译后处理
         self.translation_post_process()
@@ -216,9 +271,6 @@ class Translator(QObject, AiNieeBase):
             self.configurator.cache_list
         )
 
-        # 请求取消保存缓存文件
-        self.save_cache_file_cacelled()
-
         # 如果开启了转换简繁开关功能，则进行文本转换
         if self.configurator.response_conversion_toggle:
             with self.cache_data_lock:
@@ -263,14 +315,8 @@ class Translator(QObject, AiNieeBase):
             if self.configurator.Running_status == self.STATUS.TRANSLATION:
                 status = "翻译中"
 
-            if self.configurator.Running_status == self.STATUS.PAUSE_WAITING:
-                status = "等待暂停"
-
-            if self.configurator.Running_status == self.STATUS.PAUSE:
-                status = "已暂停"
-
-            if self.configurator.Running_status == self.STATUS.CANCEL_WAITING:
-                status = "等待取消"
+            if self.configurator.Running_status == self.STATUS.STOPING:
+                status = "停止中"
 
             # 获取结果
             result = future.result()
@@ -289,7 +335,7 @@ class Translator(QObject, AiNieeBase):
                     new["time"] = time.time() - self.data.get("start_time")
                     new["remaining_time"] = new.get("time") / max(1, new.get("line")) * new.get("remaining_line")
                     new["speed"] = new.get("total_completion_tokens") / max(1, new.get("time"))
-                    new["task"] = sum("translator" in thread.name for thread in threading.enumerate())
+                    new["task"] = len([t for t in threading.enumerate() if "translator" in t.name])
                     self.data = new
                 else:
                     new = {}
@@ -303,7 +349,7 @@ class Translator(QObject, AiNieeBase):
                     new["time"] = time.time() - self.data.get("start_time")
                     new["remaining_time"] = new.get("time") / max(1, new.get("line")) * new.get("remaining_line")
                     new["speed"] = new.get("total_completion_tokens") / max(1, new.get("time"))
-                    new["task"] = sum("translator" in thread.name for thread in threading.enumerate())
+                    new["task"] = len([t for t in threading.enumerate() if "translator" in t.name])
                     self.data = new
 
             # 更新翻译进度到缓存数据
@@ -327,36 +373,35 @@ class Translator(QObject, AiNieeBase):
     def save_cache_file_require(self):
         self.save_cache_file_require_flag = True
 
-    # 请求取消保存缓存文件
-    def save_cache_file_cacelled(self):
-        self.save_cache_file_require_flag = False
-
     # 定时保存缓存文件
     def save_cache_file_tick(self):
-        if hasattr(self, "save_cache_file_require_flag") and self.save_cache_file_require_flag == True:
-            self.save_cache_file_require_flag = False
-            with self.cache_file_lock:
-                # 创建文件夹
-                folder_path = f"{self.configurator.label_output_path}/cache"
-                os.makedirs(folder_path, exist_ok = True)
+        while True:
+            time.sleep(self.CACHE_FILE_SAVE_INTERVAL / 1000)
+            if hasattr(self, "save_cache_file_require_flag") and self.save_cache_file_require_flag == True:
+                self.save_cache_file_require_flag = False
+                with self.cache_file_lock:
+                    # 创建文件夹
+                    folder_path = f"{self.configurator.label_output_path}/cache"
+                    os.makedirs(folder_path, exist_ok = True)
 
-                # 缓存文件路径
-                file_path = f"{folder_path}/AinieeCacheData.json"
-                file_bak_path = f"{folder_path}/AinieeCacheData_bak.json"
+                    # 缓存文件路径
+                    file_path = f"{folder_path}/AinieeCacheData.json"
+                    file_bak_path = f"{folder_path}/AinieeCacheData_bak.json"
 
-                # 移除旧的备份
-                if os.path.exists(file_bak_path):
-                    os.remove(file_bak_path)
+                    # 移除旧的备份
+                    os.remove(file_bak_path) if os.path.exists(file_bak_path) else None
 
-                # 备份旧的文件
-                if os.path.exists(file_path):
-                    os.rename(file_path, file_bak_path)
+                    # 备份旧的文件
+                    os.rename(file_path, file_bak_path) if os.path.exists(file_path) else None
 
-                # 保存新的缓存文件
-                self.save_cache_file(
-                    self.configurator.cache_list,
-                    file_path
-                )
+                    # 保存新的缓存文件
+                    self.save_cache_file(
+                        self.configurator.cache_list,
+                        file_path
+                    )
+
+                    # 触发事件
+                    self.emit(self.EVENT.CACHE_FILE_AUTO_SAVE, {})
 
     # 保存缓存文件
     def save_cache_file(self, cache, path):
@@ -370,6 +415,16 @@ class Translator(QObject, AiNieeBase):
         # 写入缓存文件
         with open(path, "w", encoding = "utf-8") as writer:
             writer.write(json.dumps(new, indent = 4, ensure_ascii = False))
+
+    # 读取缓存文件
+    def load_cache_file(self, path):
+        cache = []
+
+        if os.path.exists(path):
+            with open(path, "r", encoding = "utf-8") as reader:
+                cache = json.load(reader)
+
+        return cache
 
     # 计算剩余任务总数
     def calculate_total_tasks(self, total_lines, total_tokens, lines_limit, tokens_limit, switch):
